webkitview: Change commented out code for a simple `pass`
We are already injecting code right before the `<head>` section of the HTML
file (so that we can be sure to take precedence), and we don't need another
copy potentially clashing with the objects created before (e.g., registering
newer/duplicate events).

Also, fewer lines of code, may result in a slight performance gain.

Signed-off-by: Rogério Brito <rbrito@ime.usp.br>