Update version of tunesviewer to a the 1st pre-release.
Note that, as a string, the new version (1.4.99.0) is lexicographically
smaller than the next "final" release (1.5) and, therefore, a simple
comparison with Python's `<=` works as we want.

With this small change, we can actually tag this pre-release in the git-repo
and have a saner Debian packaging (at least), especially when using
`git-buildpackage`.

It may help with other systems, like RPM too, but I don't know enough about
them to make informed assertions.

Signed-off-by: Rogério Brito <rbrito@ime.usp.br>